+++ src/Exception/KernelException.php
@@ -1,0 +1,6 @@
+<?php
+
+namespace Symlex\Exception;
+
+class KernelException extends Exception {
+}