+++ src/Kernel/Exception/ContainerNotFoundException.php
@@ -1,6 +1,0 @@
-<?php
-
-namespace Symlex\Kernel\Exception;
-
-class ContainerNotFoundException extends Exception {
-}