+++ src/Kernel/Exception/Exception.php
@@ -1,6 +1,0 @@
-<?php
-
-namespace Symlex\Kernel\Exception;
-
-class Exception extends \Exception {
-}